requirements.txt
transformers==4.55.1
huggingface_hub==0.35.3
torch>=2.0.0
pandas>=2.0.0
accelerate>=0.20.0
pyyaml>=6.0
tqdm>=4.65.0
